# docs/knowledge-base/api/openapi.json
{
  "components": {
    "schemas": {
      "AddendumCreate": {
        "properties": {
          "note": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "integer"
          },
          "target_type": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "target_type",
          "target_id",
          "note"
        ],
        "title": "AddendumCreate",
        "type": "object"
      },
      "AddendumOut": {
        "properties": {
          "author_id": {
            "title": "Author Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "integer"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "target_type",
          "target_id",
          "author_id",
          "note"
        ],
        "title": "AddendumOut",
        "type": "object"
      },
...
      "DiagnosisCreate": {
        "properties": {
          "description": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "icd_code": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Icd Code",
            "type": "string"
          }
        },
        "required": [
          "icd_code",
          "description"
        ],
        "title": "DiagnosisCreate",
        "type": "object"
      },
      "DiagnosisOut": {
        "properties": {
          "author_id": {
            "title": "Author Id",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "encounter_id": {
            "title": "Encounter Id",
            "type": "integer"
          },
          "icd_code": {
            "title": "Icd Code",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "encounter_id",
          "author_id",
          "icd_code",
          "description"
        ],
        "title": "DiagnosisOut",
        "type": "object"
      },
      "EncounterOpen": {
        "description": "Request to open an encounter from an appointment.",
        "properties": {
          "appointment_id": {
            "title": "Appointment Id",
            "type": "integer"
          }
        },
        "required": [
          "appointment_id"
        ],
        "title": "EncounterOpen",
        "type": "object"
      },
      "EncounterOut": {
        "properties": {
          "appointment_id": {
            "title": "Appointment Id",
            "type": "integer"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "consent_shared": {
            "title": "Consent Shared",
            "type": "boolean"
          },
          "doctor_id": {
            "title": "Doctor Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "integer"
          },
          "sensitive": {
            "title": "Sensitive",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "appointment_id",
          "patient_id",
          "doctor_id",
          "opened_at",
          "closed_at",
          "sensitive",
          "consent_shared"
        ],
        "title": "EncounterOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "Credentials for obtaining a token pair (story A3).",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Plaintext password to verify.",
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "PrescribeRequest": {
        "description": "A doctor's request to prescribe within an encounter (story D1).",
        "properties": {
          "dose": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Dose",
            "type": "string"
          },
          "encounter_id": {
            "title": "Encounter Id",
            "type": "integer"
          },
          "medication_id": {
            "title": "Medication Id",
            "type": "integer"
          },
          "override_interaction": {
            "default": false,
            "description": "Acknowledge and proceed past a drug-interaction warning (\u00a75.4).",
            "title": "Override Interaction",
            "type": "boolean"
          },
          "refills": {
            "default": 0,
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Refills",
            "type": "integer"
          }
        },
        "required": [
          "encounter_id",
          "medication_id",
          "dose"
        ],
        "title": "PrescribeRequest",
        "type": "object"
      },
...
      "RegisterRequest": {
        "description": "Payload for self-service patient registration (story A1).\n\nOnly patients self-register; staff accounts are admin-provisioned (A2), so\nthis request carries no role field \u2014 the service assigns the PATIENT role.",
        "properties": {
          "email": {
            "description": "Login email; must be unique.",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional display name.",
            "title": "Full Name"
          },
          "password": {
            "description": "Plaintext password (min 8 chars); hashed server-side.",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenPair": {
        "description": "The access + refresh tokens returned on login/refresh (stories A3, A4).\n\n``token_type`` is the OAuth2 bearer convention so standard clients (and the\nSwagger UI) know how to send the access token: ``Authorization: Bearer ...``.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenPair",
        "type": "object"
      },
...
      "UserOut": {
        "description": "Public representation of a user account (never includes the password hash).",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "is_active"
        ],
        "title": "UserOut",
        "type": "object"
      },
...
      "VitalsCreate": {
        "description": "Nurse-recorded vitals; every field optional (record what was measured).",
        "properties": {
          "heart_rate": {
            "anyOf": [
              {
                "maximum": 400.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate"
          },
          "resp_rate": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resp Rate"
          },
          "spo2": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spo2"
          },
          "systolic_bp": {
            "anyOf": [
              {
                "maximum": 400.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Systolic Bp"
          },
          "temp_c": {
            "anyOf": [
              {
                "maximum": 45.0,
                "minimum": 20.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp C"
          }
        },
        "title": "VitalsCreate",
        "type": "object"
      },
      "VitalsOut": {
        "properties": {
          "encounter_id": {
            "title": "Encounter Id",
            "type": "integer"
          },
          "flags": {
            "title": "Flags",
            "type": "string"
          },
          "heart_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Rate"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "recorded_by": {
            "title": "Recorded By",
            "type": "integer"
          },
          "resp_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resp Rate"
          },
          "spo2": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spo2"
          },
          "systolic_bp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Systolic Bp"
          },
          "temp_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp C"
          }
        },
        "required": [
          "id",
          "encounter_id",
          "recorded_by",
          "heart_rate",
          "resp_rate",
          "systolic_bp",
          "temp_c",
          "spo2",
          "flags"
        ],
        "title": "VitalsOut",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local-first medical portal. JSON API for accounts, appointments, clinical records, and prescriptions. The browser UI is server-rendered (Jinja2 + HTMX) by this same application.",
    "title": "HealthyVytals API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Verify credentials and return an access + refresh token pair (story A3).\n\nInvalid credentials raise ``InvalidCredentials`` \u2192 401, with an identical\nerror whether the email is unknown or the password is wrong.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log in and obtain tokens",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Self-service patient registration (story A1).\n\nReturns the created account as :class:`UserOut` (never the password hash).\nA duplicate email raises ``EmailAlreadyRegistered`` \u2192 409 via the typed-error\nhandlers.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a new patient account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/encounters": {
      "post": {
        "operationId": "open_encounter_api_v1_encounters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncounterOpen"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncounterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open an encounter (doctor only)",
        "tags": [
          "clinical"
        ]
      }
    },
    "/api/v1/encounters/addenda": {
      "post": {
        "description": "Append an immutable correction. The service restricts this to clinical staff.",
        "operationId": "add_addendum_api_v1_encounters_addenda_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddendumCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddendumOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a correction addendum (clinical staff)",
        "tags": [
          "clinical"
        ]
      }
    },
    "/api/v1/encounters/history/{patient_id}": {
      "get": {
        "description": "Return a patient's encounters, subject to treating-relationship scoping (\u00a75.3).",
        "operationId": "patient_history_api_v1_encounters_history__patient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EncounterOut"
                  },
                  "title": "Response Patient History Api V1 Encounters History  Patient Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "View a patient's encounter history (scoped)",
        "tags": [
          "clinical"
        ]
      }
    },
    "/api/v1/encounters/{encounter_id}/diagnoses": {
      "post": {
        "operationId": "add_diagnosis_api_v1_encounters__encounter_id__diagnoses_post",
        "parameters": [
          {
            "in": "path",
            "name": "encounter_id",
            "required": true,
            "schema": {
              "title": "Encounter Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a diagnosis (doctor only)",
        "tags": [
          "clinical"
        ]
      }
    },
    "/api/v1/encounters/{encounter_id}/vitals": {
      "post": {
        "operationId": "record_vitals_api_v1_encounters__encounter_id__vitals_post",
        "parameters": [
          {
            "in": "path",
            "name": "encounter_id",
            "required": true,
            "schema": {
              "title": "Encounter Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VitalsCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VitalsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record vitals (nurse only)",
        "tags": [
          "clinical"
        ]
      }
    },
...
    "/api/v1/prescriptions": {
      "post": {
        "description": "Prescribe within an encounter after \u00a75.4 safety checks (story D1).\n\nA safety block raises ``UnsafePrescription`` \u2192 409 with a ``reason`` in the\nerror details (allergy | interaction | refill_cap). An interaction can be\nproceeded past with ``override_interaction=true``; an allergy cannot.",
        "operationId": "prescribe_api_v1_prescriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prescribe a medication (doctor only, safety-checked)",
        "tags": [
          "prescriptions"
        ]
      }
    },
...
    "/api/v1/prescriptions/patient/{patient_id}": {
      "get": {
        "description": "Return a patient's prescriptions for treating staff (story D5).",
        "operationId": "patient_prescriptions_api_v1_prescriptions_patient__patient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionOut"
                  },
                  "title": "Response Patient Prescriptions Api V1 Prescriptions Patient  Patient Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A patient's prescriptions (clinical staff)",
        "tags": [
          "prescriptions"
        ]
      }
    },
...
  }
}